requirements.txt
Django==2.0
django-extensions==1.9.8
django-filter==1.1.0
djangorestframework==3.7.3
httplib2==0.10.3
oauth2==1.9.0.post1
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
typing==3.6.2
